+++ src/main/java/net/artux/pdanetwork/service/email/EmailService.java
@@ -6,7 +6,7 @@ import net.artux.pdanetwork.models.RegisterUser;
 public interface EmailService {
 
   void askForPassword(UserEntity user, String token);
-  void sendRegisterLetter(RegisterUser user, int pdaId);
+  void sendRegisterLetter(RegisterUser user, Long pdaId);
   void sendConfirmLetter(RegisterUser user, String token);
 
 }
