+++ src/main/java/net/artux/pdanetwork/service/friends/FriendService.java
@@ -1,22 +1,20 @@
 package net.artux.pdanetwork.service.friends;
 
 import net.artux.pdanetwork.models.FriendModel;
-import net.artux.pdanetwork.models.Profile;
 import net.artux.pdanetwork.models.Status;
-import net.artux.pdanetwork.models.profile.Achievement;
 
 import java.util.List;
 
 public interface FriendService {
 
   List<FriendModel> getFriends();
-  List<FriendModel> getFriends(Integer pdaId);
+  List<FriendModel> getFriends(Long pdaId);
 
   List<FriendModel> getFriendRequests();
 
   List<FriendModel> getSubs();
-  List<FriendModel> getSubs(Integer pdaId);
+  List<FriendModel> getSubs(Long pdaId);
 
-  Status addFriend(Integer pdaId);
+  Status addFriend(Long pdaId);
 
 }
